+++ galpynostatic/datasets/__init__.py
@@ -11,14 +11,14 @@
 # DOCS
 # ============================================================================
 
-r"""The ``galpynostatic.dataset`` module loads the data needed for the fits.
+r"""The ``galpynostatic.dataset`` module loads the data required for the fits.
 
-These datasets were obtained using a continuum computational physics model for
-different geometries [5]_. They come from a cutoff of multiple galvanostatic
-profiles at a given cell potential, with respect to equilibrium, with different
-combinations of the internal unitless parameters :math:`\Xi` and :math:`\ell`
-covering a wide range of possible values of the experimental descriptors
-involved.
+These datasets have been obtained using a continuum computational physics model
+for different geometries [5]_. They result from a cut-off of multiple
+galvanostatic profiles at a given cell potential, with respect to equilibrium,
+with different combinations of the internal unitless parameters :math:`\Xi` and
+:math:`\ell` covering a wide range of possible values of the experimental
+descriptors involved.
 
 References
 ----------